separated main.py into smaller modules
separated guis and functionalities in main.py into smaller modules: main_window.py and rules_tab.py
